refactor(hitcheck): split collision and damage subsystems
