+++ src/sf33rd/Source/Game/engine/hitcheck.c
@@ -563,1707 +563,7 @@ void cal_hit_mark_pos(WORK* as, WORK* ds, s16 ix2, s16 ix) {
 
 const s16 Dsas_dir_table[16] = { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0 };
 
-static void apply_head_damage_variant(PLW* as, PLW* ds) {
-    if (check_head_damage(ds->wu.routine_no[2])) {
-        ds->wu.routine_no[2] = get_kind_of_head_dm(as->wu.dir_atthit, ds->wu.dm_rl);
-    }
-}
 
-static void apply_trunk_damage_variant(PLW* as, PLW* ds) {
-    if (check_trunk_damage(ds->wu.routine_no[2])) {
-        ds->wu.routine_no[2] = get_kind_of_trunk_dm(as->wu.dir_atthit, ds->wu.dm_rl);
-    }
-}
-
-static bool uses_ground_hand_damage(s16 damage_point) {
-    if (damage_point == 4) {
-        return true;
-    }
-
-    if (damage_point == 5) {
-        return true;
-    }
-
-    if (damage_point == 6) {
-        return true;
-    }
-
-    return damage_point == 7;
-}
-
-static void resolve_ground_damage_variant(PLW* as, PLW* ds) {
-    if (ds->wu.pat_status >= 32) {
-        ds->wu.routine_no[2] = get_kagami_damage(ds->wu.routine_no[2]);
-        return;
-    }
-
-    if (ds->dm_point == 0) {
-        apply_head_damage_variant(as, ds);
-        return;
-    }
-
-    if (ds->dm_point == 1) {
-        apply_head_damage_variant(as, ds);
-        return;
-    }
-
-    if (uses_ground_hand_damage(ds->dm_point)) {
-        ds->wu.routine_no[2] = get_grd_hand_damage(ds->wu.routine_no[2]);
-    }
-
-    apply_trunk_damage_variant(as, ds);
-}
-
-static void apply_air_damage_reaction(PLW* as, PLW* ds) {
-    as->wu.hf.hit.player = 2;
-    ds->wu.kezurare_flag = 0;
-    dm_reaction_init_set(as, ds);
-
-    if (as->wu.att.dipsw & 0x10) {
-        ds->wu.routine_no[2] = get_sky_sp_damage(ds->wu.routine_no[2]);
-    } else {
-        ds->wu.routine_no[2] = get_sky_nm_damage(ds->wu.routine_no[2]);
-    }
-}
-
-static void apply_ground_damage_reaction(PLW* as, PLW* ds) {
-    as->wu.hf.hit.player = 1;
-    ds->wu.kezurare_flag = 0;
-    dm_reaction_init_set(as, ds);
-
-    if (as->wu.zu_flag == 0) {
-        resolve_ground_damage_variant(as, ds);
-    }
-}
-
-static void finish_player_damage(PLW* as, PLW* ds, s8 gddir) {
-    ds->wu.routine_no[1] = 1;
-    ds->wu.routine_no[3] = 0;
-    grade_add_clean_hits((WORK_Other*)as);
-    check_guard_miss(&as->wu, ds, gddir);
-    effect_02_init(&as->wu, ds->dm_point, 1, ds->wu.dm_rl);
-    dm_status_copy(&as->wu, &ds->wu);
-    same_dm_stop(&as->wu, &ds->wu);
-    as->wu.cmwk[8]++;
-    as->wu.cmwk[15]++;
-    ds->wu.dm_count_up++;
-
-    if (ds->wu.xyz[1].disp.pos < 0) {
-        ds->wu.xyz[1].cal = 0;
-    }
-
-    add_combo_work(as, ds);
-    hit_pattern_extdat_check(&as->wu);
-
-    if (ds->atemi_flag && ds->atemi_point != ds->dm_point) {
-        ds->atemi_flag = 0;
-    }
-
-    paring_ctr_vs[Play_Type][ds->wu.id] = 0;
-    paring_counter[ds->wu.id] = 0;
-    paring_bonus_r[ds->wu.id] = 0;
-}
-
-typedef enum {
-    PLAYER_DAMAGE_AIR_DEFENSE,
-    PLAYER_DAMAGE_GROUND_DEFENSE,
-    PLAYER_DAMAGE_AIR_REACTION,
-    PLAYER_DAMAGE_GROUND_REACTION,
-    PLAYER_DAMAGE_GUARD,
-    PLAYER_DAMAGE_PARRY,
-} PlayerDamagePath;
-
-static bool uses_forced_damage_path(PLW* as, PLW* ds) {
-    if (ds->guard_flag == 3) {
-        return true;
-    }
-
-    if (as->wu.att.guard == 0) {
-        return true;
-    }
-
-    return ds->py->flag != 0;
-}
-
-static bool uses_air_damage_path(PLW* ds) {
-    if (ds->wu.xyz[1].disp.pos > 0) {
-        return true;
-    }
-
-    return check_pat_status(&ds->wu);
-}
-
-static PlayerDamagePath select_air_or_ground_damage_path(PLW* ds, PlayerDamagePath air_path) {
-    if (uses_air_damage_path(ds)) {
-        return air_path;
-    }
-
-    return PLAYER_DAMAGE_GROUND_DEFENSE;
-}
-
-static PlayerDamagePath select_player_damage_path(PLW* as, PLW* ds) {
-    if (!uses_forced_damage_path(as, ds)) {
-        return select_air_or_ground_damage_path(ds, PLAYER_DAMAGE_AIR_DEFENSE);
-    }
-
-    if (ds->wu.pat_status == 10) {
-        ds->wu.xyz[1].cal = 0;
-        return PLAYER_DAMAGE_GROUND_DEFENSE;
-    }
-
-    if (ds->wu.pat_status == 12 && ds->wu.xyz[1].disp.pos < 6) {
-        ds->wu.xyz[1].cal = 0;
-        return PLAYER_DAMAGE_GROUND_DEFENSE;
-    }
-
-    if (ds->wu.routine_no[1] != 1) {
-        return select_air_or_ground_damage_path(ds, PLAYER_DAMAGE_AIR_DEFENSE);
-    }
-
-    if (uses_air_damage_path(ds)) {
-        return PLAYER_DAMAGE_AIR_REACTION;
-    }
-
-    return PLAYER_DAMAGE_GROUND_REACTION;
-}
-
-static PlayerDamagePath resolve_player_defense(PLW* as, PLW* ds, s8 gddir, PlayerDamagePath path) {
-    s32 defense_result;
-
-    if (path == PLAYER_DAMAGE_AIR_DEFENSE) {
-        defense_result = defense_sky(as, ds, gddir);
-    } else if (path == PLAYER_DAMAGE_GROUND_DEFENSE) {
-        defense_result = defense_ground(as, ds, gddir);
-    } else {
-        return path;
-    }
-
-    switch (defense_result) {
-    case 0:
-        return PLAYER_DAMAGE_PARRY;
-
-    case 1:
-        return PLAYER_DAMAGE_GUARD;
-
-    default:
-        return path == PLAYER_DAMAGE_AIR_DEFENSE ? PLAYER_DAMAGE_AIR_REACTION : PLAYER_DAMAGE_GROUND_REACTION;
-    }
-}
-
-void plef_at_vs_player_damage_union(PLW* as, PLW* ds, s8 gddir) { // 🟡
-    // CPS3 lacks the port-side training chip-damage metadata resets kept below.
-    PlayerDamagePath path;
-
-    ds->wu.dm_guard_success = -1;
-    path = resolve_player_defense(as, ds, gddir, select_player_damage_path(as, ds));
-
-    switch (path) {
-    case PLAYER_DAMAGE_AIR_REACTION:
-        apply_air_damage_reaction(as, ds);
-        break;
-
-    case PLAYER_DAMAGE_GROUND_REACTION:
-        apply_ground_damage_reaction(as, ds);
-        break;
-
-    case PLAYER_DAMAGE_GUARD:
-        set_guard_status(as, ds);
-        return;
-
-    case PLAYER_DAMAGE_PARRY:
-        set_paring_status(as, ds);
-        return;
-
-    default:
-        break;
-    }
-
-    finish_player_damage(as, ds, gddir);
-}
-
-static bool is_running_damage_reaction(PLW* ds) {
-    if (ds->wu.routine_no[2] == 89) {
-        return true;
-    }
-
-    return ds->wu.routine_no[2] == 90;
-}
-
-static bool should_apply_running_damage_reaction(PLW* as, PLW* ds) {
-    if (!is_running_damage_reaction(ds)) {
-        return false;
-    }
-
-    if (ds->running_f != 1) {
-        return false;
-    }
-
-    if (!Dsas_dir_table[as->wu.att.dir]) {
-        return false;
-    }
-
-    if (check_work_position(&as->wu, &ds->wu)) {
-        return ds->move_distance > 0;
-    }
-
-    return ds->move_distance < 0;
-}
-
-void dm_reaction_init_set(PLW* as, PLW* ds) { // 🟢
-    ds->wu.routine_no[2] = as->wu.att.reaction;
-
-    if (should_apply_running_damage_reaction(as, ds)) {
-        ds->wu.routine_no[2] = 99;
-    }
-
-    ds->wu.routine_no[2] = change_damage_attribute(as, as->wu.at_attribute, ds->wu.routine_no[2]);
-}
-
-void set_guard_status(PLW* as, PLW* ds) { // 🟡
-    // CPS3 always starts the guard effect; local keeps the semi-auto parry option's presentation behavior.
-    if (as->wu.att.hs_you == 0 && as->wu.att.hs_me == 0) {
-        ds->wu.routine_no[2] = ds->wu.old_rno[2];
-    } else {
-        ds->wu.routine_no[1] = 1;
-        ds->wu.routine_no[3] = 0;
-
-        if (ds->spmv_ng_flag & DIP_SEMI_AUTO_PARRY_DISABLED) {
-            effect_02_init(&as->wu, ds->dm_point, 2, ds->wu.dm_rl);
-        }
-
-        dm_status_copy(&as->wu, &ds->wu);
-        same_dm_stop(&as->wu, &ds->wu);
-
-        if (ds->wu.xyz[1].disp.pos < 0) {
-            ds->wu.xyz[1].cal = 0;
-        }
-
-        ds->wu.dm_piyo = 0;
-        as->wu.cmwk[8]++;
-        add_sp_arts_gauge_guard(as);
-        ds->wu.dm_arts_point = 0;
-        grade_add_guard_success(ds->wu.id);
-    }
-
-    hit_pattern_extdat_check(&as->wu);
-}
-
-const s8 sel_sp_ch_tbl[12] = { 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0 };
-
-const s16 sel_hs_add_tbl[6] = { 4, 3, 2, 1, 0, 0 };
-
-static s16 get_parry_hit_stop_index(PLW* as) {
-    s16 hsadix = 4;
-
-    if ((as->wu.kind_of_waza & 0xF8) == 0) {
-        hsadix = (as->wu.kind_of_waza / 2) & 3;
-    }
-
-    return hsadix;
-}
-
-static bool uses_enhanced_parry_hit_stop(s16 stop_type) {
-    if (stop_type == 0) {
-        return true;
-    }
-
-    if (stop_type == 2) {
-        return true;
-    }
-
-    return stop_type == 4;
-}
-
-static bool uses_standard_parry_hit_stop(s16 stop_type) {
-    if (stop_type == 1) {
-        return true;
-    }
-
-    if (stop_type == 3) {
-        return true;
-    }
-
-    if (stop_type == 5) {
-        return true;
-    }
-
-    if (stop_type == 6) {
-        return true;
-    }
-
-    if (stop_type == 7) {
-        return true;
-    }
-
-    if (stop_type == 8) {
-        return true;
-    }
-
-    return stop_type == 9;
-}
-
-static void apply_parry_hit_stop(PLW* as, PLW* ds, s16 hsadix) {
-    s16 stop_type = (as->wu.xyz[1].disp.pos > 0) + (ds->wu.routine_no[2] - 31) * 2;
-
-    if (uses_enhanced_parry_hit_stop(stop_type)) {
-        ds->wu.dm_stop = -15;
-        as->wu.hit_stop = sel_hs_add_tbl[hsadix] + 16;
-        as->wu.hit_quake = sel_hs_add_tbl[hsadix] + 16;
-    } else if (uses_standard_parry_hit_stop(stop_type)) {
-        ds->wu.dm_stop = -15;
-        as->wu.hit_stop = 16;
-        as->wu.hit_quake = 16;
-    } else {
-        ds->wu.dm_stop = 0;
-        as->wu.hit_stop = 0;
-        as->wu.hit_quake = 0;
-    }
-}
-
-static bool should_remake_parry_movement(PLW* as) {
-    if (as->wu.pat_status < 0xE) {
-        return false;
-    }
-
-    if (as->wu.pat_status >= 31) {
-        return false;
-    }
-
-    if (as->wu.work_id != 1) {
-        return false;
-    }
-
-    return sel_sp_ch_tbl[as->wu.kind_of_waza >> 3] == 0;
-}
-
-static void award_parry_bonus(PLW* ds) {
-    if (Bonus_Game_Flag != 0) {
-        return;
-    }
-
-    if (!(ds->spmv_ng_flag & DIP_AUTO_PARRY_DISABLED)) {
-        return;
-    }
-
-    paring_bonus_r[ds->wu.id] = 1;
-    paring_ctr_vs[Play_Type][ds->wu.id]++;
-
-    if (paring_ctr_vs[Play_Type][ds->wu.id] > 39) {
-        paring_ctr_vs[Play_Type][ds->wu.id] = 39;
-    }
-
-    paring_counter[ds->wu.id] = parisucc_pts[Play_Type][paring_ctr_vs[Play_Type][ds->wu.id] - 1];
-}
-
-static void apply_parry_status(PLW* as, PLW* ds) {
-    s16 hsadix = get_parry_hit_stop_index(as);
-
-    ds->wu.routine_no[1] = 0;
-    ds->wu.routine_no[3] = 0;
-    waza_compel_all_init2(ds);
-    dm_status_copy(&as->wu, &ds->wu);
-    ds->wu.dm_piyo = 0;
-    ds->wu.cg_type = 0;
-    apply_parry_hit_stop(as, ds, hsadix);
-    ds->wu.dm_quake = 0;
-
-    if (ds->wu.xyz[1].disp.pos < 0) {
-        ds->wu.xyz[1].cal = 0;
-    }
-
-    ds->wu.dm_arts_point = 0;
-
-    if (should_remake_parry_movement(as)) {
-        remake_mvxy_PoGR(&as->wu);
-    }
-
-    award_parry_bonus(ds);
-    as->wu.cmwk[8]++;
-}
-
-void set_paring_status(PLW* as, PLW* ds) { // 🟡
-    // CPS3 always awards parry bonuses;
-    // local's default DIP disables auto-parry, so this only differs when auto-parry is enabled.
-    if ((as->wu.att.hs_you == 0) && (as->wu.att.hs_me == 0)) {
-        ds->wu.routine_no[2] = ds->wu.old_rno[2];
-    } else {
-        apply_parry_status(as, ds);
-    }
-
-    hit_pattern_extdat_check(&as->wu);
-}
-
-s32 check_normal_attack(u8 waza) { // 🟢
-    return sel_sp_ch_tbl[waza >> 3] == 0;
-}
-
-static void apply_ground_chain_cancel(WORK* as) {
-    s16 i;
-
-    if (!(((PLW*)as)->spmv_ng_flag2 & DIP2_ALL_MOVES_CANCELLABLE_BY_HIGH_JUMP_DISABLED)) {
-        as->cg_cancel |= 1;
-    }
-
-    if (!(((PLW*)as)->spmv_ng_flag2 & DIP2_ALL_MOVES_CANCELLABLE_BY_DASH_DISABLED)) {
-        as->cg_cancel |= 2;
-    }
-
-    if (((PLW*)as)->spmv_ng_flag2 & DIP2_GROUND_CHAIN_COMBO_DISABLED) {
-        return;
-    }
-
-    i = 0;
-
-    if (((PLW*)as)->player_number == 4) {
-        as->cg_meoshi = chain_hidou_nm_ground_table[as->kow & 7];
-        as->cg_cancel |= 8;
-    } else {
-        as->cg_meoshi = i | chain_normal_ground_table[as->kow & 7];
-        as->cg_cancel |= 8;
-    }
-}
-
-static void apply_air_chain_cancel(WORK* as) {
-    s16 i;
-
-    if (((PLW*)as)->spmv_ng_flag2 & DIP2_AIR_CHAIN_COMBO_DISABLED) {
-        return;
-    }
-
-    if (hikusugi_check(as)) {
-        return;
-    }
-
-    i = 0;
-
-    if (((PLW*)as)->player_number == 7) {
-        as->cg_meoshi = chain_hidou_nm_air_table[as->kow & 7];
-        as->cg_cancel |= 8;
-    } else {
-        as->cg_meoshi = i | chain_normal_air_table[as->kow & 7];
-        as->cg_cancel |= 8;
-    }
-}
-
-static void apply_normal_move_chain_cancel(WORK* as) {
-    switch (plpat_rno_filter[as->routine_no[2]]) {
-    case 9:
-        if (as->routine_no[3] != 1) {
-            break;
-        }
-
-        /* fallthrough */
-
-    case 1:
-        apply_ground_chain_cancel(as);
-        break;
-
-    case 2:
-        apply_air_chain_cancel(as);
-        break;
-    }
-}
-
-static void apply_hit_pattern_extension(WORK* as) {
-    switch ((as->cg_extdat & 0xC0) + ((as->cg_extdat & 0x3F) != 0)) {
-    case 0x80:
-        char_move_z(as);
-        break;
-
-    case 0x40:
-        as->cg_ctr = 1;
-
-        if (!ArcadeBalance_IsEnabled()) {
-            as->cg_extdat = 0;
-        }
-
-        break;
-
-    case 0x81:
-        setup_comm_abbak(as);
-        as->cg_ix = ((as->cg_extdat & 0x3F) - 1) * as->cgd_type - as->cgd_type;
-        as->cg_next_ix = 0;
-        char_move_z(as);
-        break;
-
-    case 0x41:
-        as->cg_ctr = 1;
-        /* fallthrough */
-
-    case 0x1:
-        setup_comm_abbak(as);
-        as->cg_ix = ((as->cg_extdat & 0x3F) - 1) * as->cgd_type - as->cgd_type;
-        as->cg_next_ix = 0;
-
-        if (!ArcadeBalance_IsEnabled()) {
-            as->cg_extdat = 0;
-        }
-
-        break;
-    }
-}
-
-static void restrict_target_combo(WORK* as) {
-    if (!(((PLW*)as)->spmv_ng_flag2 & DIP2_TARGET_COMBO_DISABLED)) {
-        return;
-    }
-
-    if (!(as->cg_cancel & 8)) {
-        return;
-    }
-
-    if (as->kow & 0xF8) {
-        return;
-    }
-
-    if (as->kow & 6) {
-        as->cg_cancel &= 0xF7;
-        as->cg_meoshi = 0;
-    } else if (as->cg_meoshi & 0x110) {
-        as->cg_meoshi &= 0xF99F;
-    } else {
-        as->cg_cancel &= 0xF7;
-        as->cg_meoshi = 0;
-    }
-}
-
-static void enable_super_art_cancel(WORK* as) {
-    if (((PLW*)as)->spmv_ng_flag2 & DIP2_SA_TO_SA_CANCEL_DISABLED) {
-        return;
-    }
-
-    if (!(as->kow & 0x60)) {
-        return;
-    }
-
-    as->cg_cancel |= 0x40;
-}
-
-static void enable_special_move_cancel(WORK* as) {
-    if (((PLW*)as)->spmv_ng_flag2 & DIP2_SPECIAL_TO_SPECIAL_CANCEL_DISABLED) {
-        return;
-    }
-
-    if (as->kow & 0x60) {
-        return;
-    }
-
-    if (!(as->kow & 0xF8)) {
-        return;
-    }
-
-    as->cg_cancel |= 0x60;
-}
-
-static void enable_normal_move_cancel(WORK* as) {
-    if (((PLW*)as)->spmv_ng_flag2 & DIP2_ALL_NORMALS_CANCELLABLE_DISABLED) {
-        return;
-    }
-
-    if (as->kow & 0xF8) {
-        return;
-    }
-
-    switch (plpat_rno_filter[as->routine_no[2]]) {
-    case 9:
-        if (as->routine_no[3] != 1) {
-            break;
-        }
-
-        /* fallthrough */
-
-    case 1:
-    case 2:
-        as->cg_cancel |= 0x60;
-        break;
-    }
-}
-
-static bool can_apply_normal_move_chain(WORK* as) {
-    if (as->kow & 0xF8) {
-        return false;
-    }
-
-    if (as->routine_no[1] != 4) {
-        return false;
-    }
-
-    return as->routine_no[2] < 0x10;
-}
-
-void hit_pattern_extdat_check(WORK* as) { // 🟡
-    // CPS3 leaves cg_extdat latched; local's extra cancel rewrites are DIP-gated and default off.
-    apply_hit_pattern_extension(as);
-
-    if (as->work_id != 1) {
-        return;
-    }
-
-    restrict_target_combo(as);
-    enable_super_art_cancel(as);
-    enable_special_move_cancel(as);
-    enable_normal_move_cancel(as);
-
-    if (can_apply_normal_move_chain(as)) {
-        apply_normal_move_chain_cancel(as);
-    }
-}
-
-static bool prevents_chip_damage_ko(WORK* as, s16 player_id) {
-    return as->no_death_attack || (plw[player_id].spmv_ng_flag2 & DIP2_CHIP_DAMAGE_KO_DISABLED);
-}
-
-static s16 damage_attacker_player_id(WORK* attacker) {
-    if (attacker->work_id == 1) {
-        return attacker->id;
-    }
-
-    return ((WORK_Other*)attacker)->master_id;
-}
-
-s16 check_dm_att_guard(WORK* as, WORK* ds, s16 kom) { // 🟡
-    // CPS3 uses raw kezuri_pow; local defaults match CPS3 for chip DIPs, while non-arcade keeps the kom scaling.
-    s16 curr_id;
-    s16 divisor;
-    s16 rnum;
-
-    rnum = 0;
-    ds->kezurare_flag = 0;
-    curr_id = damage_attacker_player_id(as);
-
-    if (!(plw[curr_id].spmv_ng_flag & DIP_CHIP_DAMAGE_ENABLED)) {
-        as->kezuri_pow = 0;
-    }
-
-    if (!as->kezuri_pow) {
-        ds->dm_vital = 0;
-        return rnum;
-    }
-
-    if (ds->dm_vital == 0) {
-        return rnum;
-    }
-
-    ds->kezurare_flag = 1;
-    divisor = as->kezuri_pow;
-
-    if (!ArcadeBalance_IsEnabled()) {
-        divisor /= kom;
-    }
-
-    ds->dm_vital = ds->dm_vital / divisor;
-
-    if (ds->dm_vital == 0) {
-        ds->dm_vital = 1;
-    }
-
-    if (ds->dm_vital <= ds->vital_new) {
-        return rnum;
-    }
-
-    if (prevents_chip_damage_ko(as, curr_id)) {
-        ds->dm_vital = ds->vital_new;
-    } else {
-        ds->dm_guard_success = ds->routine_no[2];
-        rnum = 1;
-    }
-
-    return rnum;
-}
-
-static bool supports_blocking_chip_damage(WORK* as, TAMA* tama) {
-    if (as->work_id != 4) {
-        return false;
-    }
-
-    if (as->id != 13) {
-        return false;
-    }
-
-    if (tama->kz_blocking == 0) {
-        return false;
-    }
-
-    return as->kezuri_pow;
-}
-
-s16 check_dm_att_blocking(WORK* as, WORK* ds, s16 dnum) { // 🟡
-    // CPS3 lacks the port-side training chip-damage metadata kept below.
-    s16 rnum = 0;
-    TAMA* tama = (TAMA*)as->my_effadrs;
-
-    ds->kezurare_flag = 0;
-
-    if (!supports_blocking_chip_damage(as, tama)) {
-        ds->dm_vital = 0;
-        return rnum;
-    }
-
-    if (ds->dm_vital == 0) {
-        return rnum;
-    }
-
-    ds->kezurare_flag = 1;
-
-    if (as->kezuri_pow) {
-        ds->dm_vital = ds->dm_vital / as->kezuri_pow;
-    } else {
-        ds->dm_vital = 0;
-    }
-
-    if (ds->dm_vital == 0) {
-        ds->dm_vital = 1;
-    }
-
-    if (ds->dm_vital <= ds->vital_new) {
-        return rnum;
-    }
-
-    if (as->no_death_attack) {
-        ds->dm_vital = ds->vital_new;
-    } else {
-        ds->dm_guard_success = dnum;
-        rnum = 1;
-    }
-
-    return rnum;
-}
-
-static bool uses_medium_damage_modifier(s16 status) {
-    if (status == 32) {
-        return true;
-    }
-
-    if (status == 3) {
-        return true;
-    }
-
-    return status == 25;
-}
-
-static bool uses_high_damage_modifier(s16 status) {
-    if (status == 7) {
-        return true;
-    }
-
-    if (status == 23) {
-        return true;
-    }
-
-    return status == 35;
-}
-
-static bool uses_double_damage_modifier(s16 status) {
-    if (status == 1) {
-        return true;
-    }
-
-    if (status == 21) {
-        return true;
-    }
-
-    return status == 37;
-}
-
-static void apply_status_damage_modifier(PLW* ds) {
-    if (uses_medium_damage_modifier(ds->wu.pat_status)) {
-        ds->wu.dm_vital = (ds->wu.dm_vital * 125) / 100;
-    } else if (uses_high_damage_modifier(ds->wu.pat_status)) {
-        ds->wu.dm_vital = (ds->wu.dm_vital * 150) / 100;
-    } else if (uses_double_damage_modifier(ds->wu.pat_status)) {
-        ds->wu.dm_vital *= 2;
-    }
-}
-
-static void apply_throw_damage_scaling(PLW* as, PLW* ds) {
-    ds->wu.dm_vital = (ds->wu.dm_vital) * (as->tk_nage + 32) / 32;
-
-    if ((as->tk_nage -= 2) < 0) {
-        as->tk_nage = 0;
-    }
-}
-
-static void apply_strike_damage_scaling(PLW* as, PLW* ds) {
-    ds->wu.dm_vital = (ds->wu.dm_vital) * (as->tk_dageki + 32) / 32;
-
-    if ((as->tk_dageki -= 2) < 0) {
-        as->tk_dageki = 0;
-    }
-}
-
-static void apply_attack_damage_scaling(PLW* as, PLW* ds) {
-    if (!ds->wu.dm_vital) {
-        return;
-    }
-
-    if (as->wu.routine_no[1] == 2) {
-        apply_throw_damage_scaling(as, ds);
-    }
-
-    if (as->wu.routine_no[1] == 4) {
-        apply_strike_damage_scaling(as, ds);
-    }
-
-    ds->utk_nage = as->tk_nage;
-    ds->utk_dageki = as->tk_dageki;
-}
-
-static void apply_stun_scaling(PLW* as, PLW* ds) {
-    if (!ds->wu.dm_piyo) {
-        return;
-    }
-
-    ds->wu.dm_piyo = ds->wu.dm_piyo * (as->tk_kizetsu + 32) / 32;
-
-    if ((as->tk_kizetsu -= 2) < 0) {
-        as->tk_kizetsu = 0;
-    }
-
-    ds->utk_kizetsu = as->tk_kizetsu;
-}
-
-static void apply_damage_vital_backup(PLW* as, PLW* ds) {
-    switch (as->dm_vital_use) {
-    case 1:
-        ds->wu.dm_vital += as->dm_vital_backup;
-        as->dm_vital_backup = 0;
-        break;
-
-    case 2:
-        as->dm_vital_backup /= 2;
-        ds->wu.dm_vital += as->dm_vital_backup;
-        break;
-    }
-}
-
-void set_damage_and_piyo(PLW* as, PLW* ds) { // 🟡
-    // CPS3 has fixed stun gain; local's extra-option stun multiplier is neutral by default.
-    cal_damage_vitality(as, ds);
-    ds->wu.dm_piyo = _add_piyo_gauge[as->player_number][as->wu.att.piyo];
-    ds->wu.dm_piyo = ds->wu.dm_piyo * stun_gauge_omake[omop_stun_gauge_add[(ds->wu.id + 1) & 1]] / 32;
-    apply_status_damage_modifier(ds);
-    apply_attack_damage_scaling(as, ds);
-    apply_stun_scaling(as, ds);
-    as->wu.at_ten_ix = remake_score_index(ds->wu.dm_vital);
-    cal_combo_waribiki(as, ds);
-    cal_dm_vital_gauge_hosei(ds);
-    cal_combo_waribiki2(ds);
-
-    if (as->wu.work_id != 1) {
-        return;
-    }
-
-    apply_damage_vital_backup(as, ds);
-}
-
-s16 remake_score_index(s16 dmv) { // 🟢
-    s16 i;
-
-    for (i = 0; i < 16; i++) {
-        if (dmv < rsix_r_table[i][0]) {
-            break;
-        }
-    }
-
-    return rsix_r_table[i][1];
-}
-
-void same_dm_stop(WORK* as, WORK* ds) { // 🟢
-    if (as->work_id != 1) {
-        return;
-    }
-
-    if (!(as->att.dipsw & 1)) {
-        return;
-    }
-
-    if (ds->xyz[1].disp.pos <= 0 && (ds->vital_new - ds->dm_vital) >= -2) {
-        return;
-    }
-
-    switch ((ds->dm_stop < 0) + ((as->att.hs_me < 0) * 2)) {
-    case 1:
-        ds->dm_stop = -as->att.hs_me;
-        /* fallthrough */
-
-    case 2:
-        ds->dm_stop = -as->att.hs_me;
-        break;
-
-    default:
-        ds->dm_stop = as->att.hs_me;
-        break;
-    }
-}
-
-static bool can_attempt_cps3_air_parry(PLW* as, PLW* ds) {
-    if (ds->py->flag != 0) {
-        return false;
-    }
-
-    if (ds->guard_flag & 2) {
-        return false;
-    }
-
-    return as->wu.att.guard & 4;
-}
-
-static bool can_cps3_air_parry(PLW* ds) {
-    if (ds->spmv_ng_flag & DIP_AIR_PARRY_DISABLED) {
-        return false;
-    }
-
-    return ds->cp->waza_flag[5] != 0;
-}
-
-static bool can_cps3_anti_air_parry(PLW* ds) {
-    if (ds->spmv_ng_flag & DIP_ANTI_AIR_PARRY_DISABLED) {
-        return false;
-    }
-
-    return ds->cp->waza_flag[6] != 0;
-}
-
-static s32 resolve_cps3_air_parry(PLW* as, PLW* ds, s16 reaction) {
-    blocking_point_count_up(ds);
-    as->wu.hf.hit.player = 0x80;
-    ds->wu.routine_no[2] = reaction;
-
-    if (check_dm_att_blocking(&as->wu, &ds->wu, 7)) {
-        return 2;
-    }
-
-    return 0;
-}
-
-static bool is_cps3_air_guard_rejected(PLW* as, PLW* ds, s8 gddir) {
-    if (!(as->wu.att.guard & 0x20)) {
-        return true;
-    }
-
-    if (ds->guard_flag & 1) {
-        return true;
-    }
-
-    if (ds->spmv_ng_flag & DIP_AIR_GUARD_DISABLED) {
-        return true;
-    }
-
-    return !(gddir & ds->saishin_lvdir);
-}
-
-static s32 finish_cps3_air_guard(PLW* as, PLW* ds) {
-    as->wu.hf.hit.player = 0x20;
-    ds->wu.routine_no[2] = 7;
-
-    if (check_dm_att_guard(&as->wu, &ds->wu, 2)) {
-        return 2;
-    }
-
-    return 1;
-}
-
-s32 defense_sky_cps3(PLW* as, PLW* ds, s8 gddir) { // 🟢
-    if (can_attempt_cps3_air_parry(as, ds)) {
-        if (can_cps3_air_parry(ds)) {
-            return resolve_cps3_air_parry(as, ds, 0x22);
-        }
-
-        if (can_cps3_anti_air_parry(ds)) {
-            return resolve_cps3_air_parry(as, ds, 0x23);
-        }
-    }
-
-    if (is_cps3_air_guard_rejected(as, ds, gddir)) {
-        return 2;
-    }
-
-    return finish_cps3_air_guard(as, ds);
-}
-
-typedef struct {
-    PLW* attacker;
-    PLW* defender;
-    s8 just_now;
-    s8 attack_attribute;
-    s8 automatic_parry;
-    s8 automatic_guard;
-} Ps2AirDefense;
-
-static bool can_ps2_red_air_parry(const Ps2AirDefense* defense) {
-    if (defense->defender->spmv_ng_flag & DIP_RED_PARRY_DISABLED) {
-        return false;
-    }
-
-    if (defense->defender->cp->waza_flag[5] >=
-        grdb2[defense->defender->wu.id][defense->attack_attribute]) {
-        return true;
-    }
-
-    return defense->automatic_parry;
-}
-
-static bool can_ps2_air_parry(const Ps2AirDefense* defense) {
-    if (defense->defender->py->flag != 0) {
-        return false;
-    }
-
-    if (defense->defender->guard_flag & 2) {
-        return false;
-    }
-
-    if (!(defense->attacker->wu.att.guard & 4)) {
-        return false;
-    }
-
-    if (defense->just_now) {
-        return can_ps2_red_air_parry(defense);
-    }
-
-    if (defense->defender->spmv_ng_flag & DIP_AIR_PARRY_DISABLED) {
-        return false;
-    }
-
-    if (defense->defender->cp->waza_flag[5] != 0) {
-        return true;
-    }
-
-    return defense->automatic_parry;
-}
-
-static s32 resolve_ps2_air_parry(const Ps2AirDefense* defense) {
-    blocking_point_count_up(defense->defender);
-    defense->attacker->wu.hf.hit.player = 0x80;
-    defense->defender->wu.routine_no[2] = 0x22;
-
-    if (check_dm_att_blocking(&defense->attacker->wu, &defense->defender->wu, 7)) {
-        return 2;
-    }
-
-    return 0;
-}
-
-static bool bypasses_ps2_air_guard_input(const Ps2AirDefense* defense) {
-    if (defense->defender->auto_guard) {
-        return true;
-    }
-
-    if (defense->automatic_guard) {
-        return true;
-    }
-
-    if (defense->defender->spmv_ng_flag & DIP_ABSOLUTE_GUARD_DISABLED) {
-        return false;
-    }
-
-    return defense->just_now;
-}
-
-static bool is_ps2_air_guard_rejected(const Ps2AirDefense* defense, s8 gddir) {
-    if (!(defense->attacker->wu.att.guard & 32)) {
-        return true;
-    }
-
-    if (defense->defender->guard_flag & 1) {
-        return true;
-    }
-
-    if (defense->defender->spmv_ng_flag & DIP_AIR_GUARD_DISABLED) {
-        return true;
-    }
-
-    if (bypasses_ps2_air_guard_input(defense)) {
-        return false;
-    }
-
-    if (!(defense->defender->saishin_lvdir & gddir)) {
-        return true;
-    }
-
-    return defense->defender->cp->sw_lvbt & 3;
-}
-
-static s32 finish_ps2_air_guard(const Ps2AirDefense* defense) {
-    defense->attacker->wu.hf.hit.player = 0x20;
-    defense->defender->wu.routine_no[2] = 7;
-
-    if (check_dm_att_guard(&defense->attacker->wu, &defense->defender->wu, 2)) {
-        return 2;
-    }
-
-    return 1;
-}
-
-s32 defense_sky_ps2(PLW* as, PLW* ds, s8 gddir) { // 🔴
-    Ps2AirDefense defense;
-
-    defense.attacker = as;
-    defense.defender = ds;
-    defense.automatic_parry = (ds->spmv_ng_flag & DIP_AUTO_PARRY_DISABLED) == 0;
-    defense.automatic_guard = (ds->spmv_ng_flag & DIP_AUTO_GUARD_DISABLED) == 0;
-
-    if (ds->dead_flag) {
-        ds->guard_flag = 3;
-    }
-
-    defense.just_now = 0;
-
-    if (ds->guard_chuu != 0 && ds->guard_chuu < 5) {
-        defense.just_now = 1;
-        defense.attack_attribute = check_normal_attack(as->wu.kind_of_waza);
-    }
-
-    if (can_ps2_air_parry(&defense)) {
-        return resolve_ps2_air_parry(&defense);
-    }
-
-    if (is_ps2_air_guard_rejected(&defense, gddir)) {
-        return 2;
-    }
-
-    return finish_ps2_air_guard(&defense);
-}
-
-s32 defense_sky(PLW* as, PLW* ds, s8 gddir) { // 🟡
-    // CPS3 lacks the local red-parry/auto-guard extensions and has a second air-parry branch.
-    if (ArcadeBalance_IsEnabled()) {
-        return defense_sky_cps3(as, ds, gddir);
-    } else {
-        return defense_sky_ps2(as, ds, gddir);
-    }
-}
-
-static bool is_standing_blocking_state(PLW* wk) {
-    if (wk->wu.routine_no[1] == 0) {
-        if (wk->wu.routine_no[2] > 30) {
-            return wk->wu.routine_no[2] < 36;
-        }
-    }
-
-    return false;
-}
-
-static bool is_crouching_blocking_state(PLW* wk) {
-    if (wk->wu.routine_no[1] == 1) {
-        if (wk->wu.routine_no[2] > 3) {
-            return wk->wu.routine_no[2] < 8;
-        }
-    }
-
-    return false;
-}
-
-void blocking_point_count_up(PLW* wk) { // 🟡
-    // CPS3 always scores parries; local's default DIP disables auto-parry, so this only differs when auto-parry is
-    // enabled.
-    wk->kind_of_blocking = 0;
-
-    if (is_standing_blocking_state(wk)) {
-        wk->kind_of_blocking = 1;
-    }
-
-    if (is_crouching_blocking_state(wk)) {
-        wk->kind_of_blocking = 2;
-    }
-
-    if (wk->spmv_ng_flag & DIP_AUTO_PARRY_DISABLED) {
-        grade_add_blocking(wk);
-    }
-}
-
-static s32 resolve_standing_guard_block_cps3(PLW* as, PLW* ds) {
-    blocking_point_count_up(ds);
-    as->wu.hf.hit.player = 0x40;
-
-    if (check_attbox_dir(ds) == 0) {
-        ds->wu.routine_no[2] = 0x1F;
-    } else {
-        ds->wu.routine_no[2] = 0x20;
-    }
-
-    if (check_dm_att_blocking(&as->wu, &ds->wu, 5)) {
-        return 2;
-    }
-
-    return 0;
-}
-
-static s32 resolve_crouching_guard_block_cps3(PLW* as, PLW* ds) {
-    blocking_point_count_up(ds);
-    as->wu.hf.hit.player = 0x40;
-    ds->wu.routine_no[2] = 0x21;
-
-    if (check_dm_att_blocking(&as->wu, &ds->wu, 6)) {
-        return 2;
-    }
-
-    return 0;
-}
-
-typedef struct {
-    PLW* attacker;
-    PLW* defender;
-    s8 just_now;
-    s8 attack_attribute;
-} Cps3GroundDefense;
-
-static bool can_cps3_standing_parry(const Cps3GroundDefense* defense) {
-    if (defense->just_now) {
-        return defense->defender->cp->waza_flag[3] >=
-               grdb[defense->defender->wu.id][defense->attack_attribute][0];
-    }
-
-    if (defense->attacker->wu.jump_att_flag == 0) {
-        return defense->defender->cp->waza_flag[3] != 0;
-    }
-
-    return defense->defender->cp->waza_flag[12] != 0;
-}
-
-static bool try_cps3_standing_parry(const Cps3GroundDefense* defense, s32* result) {
-    if (!(defense->attacker->wu.att.guard & 2)) {
-        return false;
-    }
-
-    if (defense->defender->spmv_ng_flag & DIP_UNKNOWN_8) {
-        return false;
-    }
-
-    if (!can_cps3_standing_parry(defense)) {
-        return false;
-    }
-
-    *result = resolve_standing_guard_block_cps3(defense->attacker, defense->defender);
-    return true;
-}
-
-static bool can_cps3_crouching_parry(const Cps3GroundDefense* defense) {
-    if (defense->just_now) {
-        return defense->defender->cp->waza_flag[4] >=
-               grdb[defense->defender->wu.id][defense->attack_attribute][1];
-    }
-
-    return defense->defender->cp->waza_flag[4] != 0;
-}
-
-static bool try_cps3_crouching_parry(const Cps3GroundDefense* defense, s32* result) {
-    if (!(defense->attacker->wu.att.guard & 1)) {
-        return false;
-    }
-
-    if (defense->defender->spmv_ng_flag & DIP_UNKNOWN_9) {
-        return false;
-    }
-
-    if (!can_cps3_crouching_parry(defense)) {
-        return false;
-    }
-
-    *result = resolve_crouching_guard_block_cps3(defense->attacker, defense->defender);
-    return true;
-}
-
-static bool try_cps3_ground_parry(const Cps3GroundDefense* defense, s32* result) {
-    if (defense->defender->py->flag != 0) {
-        return false;
-    }
-
-    if (defense->defender->guard_flag & 2) {
-        return false;
-    }
-
-    if (!(defense->attacker->wu.att.guard & 3)) {
-        return false;
-    }
-
-    if (try_cps3_standing_parry(defense, result)) {
-        return true;
-    }
-
-    return try_cps3_crouching_parry(defense, result);
-}
-
-static bool is_cps3_ground_guard_rejected(const Cps3GroundDefense* defense, s8 gddir) {
-    if (!(defense->attacker->wu.att.guard & 0x18)) {
-        return true;
-    }
-
-    if (defense->defender->guard_flag & 1) {
-        return true;
-    }
-
-    if (defense->defender->spmv_ng_flag & DIP_GUARD_DISABLED) {
-        return true;
-    }
-
-    if (defense->defender->auto_guard) {
-        return false;
-    }
-
-    if (!(defense->defender->saishin_lvdir & gddir)) {
-        return true;
-    }
-
-    return defense->defender->cp->sw_lvbt & 1;
-}
-
-static bool configure_cps3_guard_animation(const Cps3GroundDefense* defense) {
-    switch (defense->attacker->wu.att.guard & 0x18) {
-    case 8:
-        if (!(defense->defender->cp->sw_lvbt & 2)) {
-            return false;
-        }
-
-        defense->defender->wu.routine_no[2] = 6;
-        return true;
-
-    case 16:
-        if (defense->defender->cp->sw_lvbt & 2) {
-            return false;
-        }
-
-        defense->defender->wu.routine_no[2] = 5;
-        return true;
-
-    default:
-        if (defense->defender->cp->sw_lvbt & 2) {
-            defense->defender->wu.routine_no[2] = 6;
-        } else {
-            defense->defender->wu.routine_no[2] = 5;
-        }
-
-        return true;
-    }
-}
-
-static s32 finish_cps3_ground_guard(const Cps3GroundDefense* defense) {
-    defense->attacker->wu.hf.hit.player = 0x10;
-
-    if (defense->defender->wu.routine_no[2] == 5 && check_attbox_dir(defense->defender) == 0) {
-        defense->defender->wu.routine_no[2] = 4;
-    }
-
-    if (check_dm_att_guard(&defense->attacker->wu, &defense->defender->wu, 1)) {
-        return 2;
-    }
-
-    return 1;
-}
-
-s32 defense_ground_cps3(PLW* as, PLW* ds, s8 gddir) { // 🟢
-    Cps3GroundDefense defense;
-    s32 parry_result;
-
-    defense.attacker = as;
-    defense.defender = ds;
-    defense.just_now = 0;
-    defense.attack_attribute = 0;
-
-    if (ds->guard_chuu != 0 && ds->guard_chuu < 5) {
-        defense.just_now = 1;
-        defense.attack_attribute = check_normal_attack(as->wu.kind_of_waza);
-    }
-
-    if (try_cps3_ground_parry(&defense, &parry_result)) {
-        return parry_result;
-    }
-
-    if (is_cps3_ground_guard_rejected(&defense, gddir)) {
-        return 2;
-    }
-
-    if (!configure_cps3_guard_animation(&defense)) {
-        return 2;
-    }
-
-    return finish_cps3_ground_guard(&defense);
-}
-
-static s32 resolve_standing_guard_block(PLW* as, PLW* ds) {
-    blocking_point_count_up(ds);
-    as->wu.hf.hit.player = 64;
-
-    if (check_attbox_dir(ds) == 0) {
-        ds->wu.routine_no[2] = 31;
-    } else {
-        ds->wu.routine_no[2] = 32;
-    }
-
-    if (check_dm_att_blocking(&as->wu, &ds->wu, 5)) {
-        return 2;
-    }
-
-    return 0;
-}
-
-static s32 resolve_crouching_guard_block(PLW* as, PLW* ds) {
-    blocking_point_count_up(ds);
-    as->wu.hf.hit.player = 64;
-    ds->wu.routine_no[2] = 33;
-
-    if (check_dm_att_blocking(&as->wu, &ds->wu, 6)) {
-        return 2;
-    }
-
-    return 0;
-}
-
-typedef struct {
-    PLW* attacker;
-    PLW* defender;
-    s8 just_now;
-    s8 attack_attribute;
-    s8 automatic_parry;
-    s8 automatic_guard;
-} Ps2GroundDefense;
-
-typedef enum {
-    PARRY_STANDING,
-    PARRY_CROUCHING,
-} ParryStance;
-
-static bool attack_supports_parry_stance(const Ps2GroundDefense* defense, ParryStance stance) {
-    if (stance == PARRY_STANDING) {
-        return defense->attacker->wu.att.guard & 2;
-    }
-
-    return defense->attacker->wu.att.guard & 1;
-}
-
-static bool can_red_parry(const Ps2GroundDefense* defense, ParryStance stance) {
-    if (defense->defender->spmv_ng_flag & DIP_RED_PARRY_DISABLED) {
-        return false;
-    }
-
-    if (stance == PARRY_STANDING) {
-        return (defense->defender->cp->waza_flag[3] >=
-                grdb[defense->defender->wu.id][defense->attack_attribute][0]) ||
-               defense->automatic_parry;
-    }
-
-    return !(defense->defender->cp->waza_flag[4] <
-             grdb[defense->defender->wu.id][defense->attack_attribute][1]) ||
-           defense->automatic_parry;
-}
-
-static bool is_standard_parry_disabled(const Ps2GroundDefense* defense, ParryStance stance) {
-    if (stance == PARRY_STANDING) {
-        return defense->defender->spmv_ng_flag & DIP_UNKNOWN_8;
-    }
-
-    return defense->defender->spmv_ng_flag & DIP_UNKNOWN_9;
-}
-
-static bool can_air_parry(const Ps2GroundDefense* defense, ParryStance stance) {
-    if (defense->defender->spmv_ng_flag & DIP_ANTI_AIR_PARRY_DISABLED) {
-        return false;
-    }
-
-    if (stance == PARRY_STANDING) {
-        return defense->defender->cp->waza_flag[12] != 0 || defense->automatic_parry;
-    }
-
-    return defense->defender->cp->waza_flag[4] != 0 || defense->automatic_parry;
-}
-
-static bool can_ground_parry(const Ps2GroundDefense* defense, ParryStance stance) {
-    if (stance == PARRY_STANDING) {
-        return defense->defender->cp->waza_flag[3] != 0 || defense->automatic_parry;
-    }
-
-    return defense->defender->cp->waza_flag[4] != 0 || defense->automatic_parry;
-}
-
-static bool can_standard_parry(const Ps2GroundDefense* defense, ParryStance stance) {
-    if (is_standard_parry_disabled(defense, stance)) {
-        return false;
-    }
-
-    if (defense->attacker->wu.jump_att_flag) {
-        return can_air_parry(defense, stance);
-    }
-
-    return can_ground_parry(defense, stance);
-}
-
-static bool try_ps2_parry(const Ps2GroundDefense* defense, ParryStance stance, s32* result) {
-    bool can_parry;
-
-    if (!attack_supports_parry_stance(defense, stance)) {
-        return false;
-    }
-
-    if (defense->just_now) {
-        can_parry = can_red_parry(defense, stance);
-    } else {
-        can_parry = can_standard_parry(defense, stance);
-    }
-
-    if (!can_parry) {
-        return false;
-    }
-
-    if (stance == PARRY_STANDING) {
-        *result = resolve_standing_guard_block(defense->attacker, defense->defender);
-    } else {
-        *result = resolve_crouching_guard_block(defense->attacker, defense->defender);
-    }
-
-    return true;
-}
-
-static bool try_ps2_ground_parry(const Ps2GroundDefense* defense, s32* result) {
-    if (defense->defender->py->flag != 0) {
-        return false;
-    }
-
-    if (defense->defender->guard_flag & 2) {
-        return false;
-    }
-
-    if (!(defense->attacker->wu.att.guard & 3)) {
-        return false;
-    }
-
-    if (try_ps2_parry(defense, PARRY_STANDING, result)) {
-        return true;
-    }
-
-    return try_ps2_parry(defense, PARRY_CROUCHING, result);
-}
-
-static bool bypasses_manual_guard_input(const Ps2GroundDefense* defense) {
-    if (defense->defender->auto_guard) {
-        return true;
-    }
-
-    if (defense->automatic_guard) {
-        return true;
-    }
-
-    if (defense->defender->spmv_ng_flag & DIP_ABSOLUTE_GUARD_DISABLED) {
-        return false;
-    }
-
-    return defense->just_now;
-}
-
-static bool is_manual_guard_input_rejected(const Ps2GroundDefense* defense, s8 gddir) {
-    if (bypasses_manual_guard_input(defense)) {
-        return false;
-    }
-
-    if (!(defense->defender->saishin_lvdir & gddir)) {
-        return true;
-    }
-
-    return defense->defender->cp->sw_lvbt & 1;
-}
-
-static bool is_ps2_ground_guard_rejected(const Ps2GroundDefense* defense, s8 gddir) {
-    if (!(defense->attacker->wu.att.guard & 0x18)) {
-        return true;
-    }
-
-    if (defense->defender->guard_flag & 1) {
-        return true;
-    }
-
-    if (defense->defender->spmv_ng_flag & DIP_GUARD_DISABLED) {
-        return true;
-    }
-
-    return is_manual_guard_input_rejected(defense, gddir);
-}
-
-static bool configure_directional_guard_animation(const Ps2GroundDefense* defense, bool requires_down,
-                                                  s16 animation) {
-    bool is_down = defense->defender->cp->sw_lvbt & 2;
-
-    if (defense->automatic_guard == 0 && is_down != requires_down) {
-        return false;
-    }
-
-    defense->defender->wu.routine_no[2] = animation;
-    return true;
-}
-
-static bool configure_ps2_guard_animation(const Ps2GroundDefense* defense) {
-    switch (defense->attacker->wu.att.guard & 0x18) {
-    case 8:
-        return configure_directional_guard_animation(defense, true, 6);
-
-    case 16:
-        return configure_directional_guard_animation(defense, false, 5);
-
-    default:
-        if (defense->defender->cp->sw_lvbt & 2) {
-            defense->defender->wu.routine_no[2] = 6;
-        } else {
-            defense->defender->wu.routine_no[2] = 5;
-        }
-
-        return true;
-    }
-}
-
-static s32 finish_ps2_ground_guard(const Ps2GroundDefense* defense) {
-    defense->attacker->wu.hf.hit.player = 16;
-
-    if (defense->defender->wu.routine_no[2] == 5 && check_attbox_dir(defense->defender) == 0) {
-        defense->defender->wu.routine_no[2] = 4;
-    }
-
-    if (check_dm_att_guard(&defense->attacker->wu, &defense->defender->wu, 1)) {
-        return 2;
-    }
-
-    return 1;
-}
-
-s32 defense_ground_ps2(PLW* as, PLW* ds, s8 gddir) { // 🔴
-    Ps2GroundDefense defense;
-    s32 parry_result;
-
-    defense.attacker = as;
-    defense.defender = ds;
-    defense.automatic_parry = (ds->spmv_ng_flag & DIP_AUTO_PARRY_DISABLED) == 0;
-    defense.automatic_guard = (ds->spmv_ng_flag & DIP_AUTO_GUARD_DISABLED) == 0;
-
-    if (ds->dead_flag) {
-        ds->guard_flag = 3;
-    }
-
-    defense.just_now = 0;
-
-    if (ds->guard_chuu != 0 && ds->guard_chuu < 5) {
-        defense.just_now = 1;
-        defense.attack_attribute = check_normal_attack(as->wu.kind_of_waza);
-    }
-
-    if (try_ps2_ground_parry(&defense, &parry_result)) {
-        return parry_result;
-    }
-
-    if (is_ps2_ground_guard_rejected(&defense, gddir)) {
-        return 2;
-    }
-
-    if (!configure_ps2_guard_animation(&defense)) {
-        return 2;
-    }
-
-    return finish_ps2_ground_guard(&defense);
-}
-
-s32 defense_ground(PLW* as, PLW* ds, s8 gddir) { // 🟡
-    // CPS3 lacks the local red-parry/auto-guard extensions in the PS2 ground-defense path.
-    if (ArcadeBalance_IsEnabled()) {
-        return defense_ground_cps3(as, ds, gddir);
-    } else {
-        return defense_ground_ps2(as, ds, gddir);
-    }
-}
 
 static bool preserves_attack_direction(WORK* as, WORK* ds) {
     if (as->work_id != 1) {
@@ -2482,662 +782,6 @@ void cal_combo_waribiki2(PLW* ds) { // 🟢
     }
 }
 
-static bool should_skip_normal_catch(WORK* target) {
-    return (target->routine_no[1] == 1) && (target->routine_no[3] != 0) && (target->routine_no[2] != 0x19);
-}
-
-static bool should_skip_oiuchi_catch(WORK* target) {
-    return (target->routine_no[1] == 1) && (target->routine_no[3] != 0) && (target->cg_type != 10) &&
-           !dm_oiuchi_catch[target->routine_no[2]];
-}
-
-static bool should_skip_catch_target(WORK* mad, WORK* sad) {
-    if (mad->att.guard & 0x18) {
-        return false;
-    }
-
-    if (((PLW*)sad)->tsukamarenai_flag) {
-        return true;
-    }
-
-    if (!(mad->att.dipsw & 0x60)) {
-        return should_skip_normal_catch(sad);
-    }
-
-    return should_skip_oiuchi_catch(sad);
-}
-
-typedef struct {
-    WORK* attacker;
-    WORK* target;
-    s16* attack_box;
-    s16* target_box;
-    s16 attacker_index;
-    s16 target_index;
-} CatchScan;
-
-static bool prepare_catch_attacker(CatchScan* scan, s16 attacker_index) {
-    scan->attacker_index = attacker_index;
-
-    if (hs[attacker_index].flag.results & 0x1000) {
-        return false;
-    }
-
-    scan->attacker = q_hit_push[attacker_index];
-
-    if (scan->attacker->work_id != 1) {
-        return false;
-    }
-
-    if (scan->attacker->att_hit_ok == 0) {
-        return false;
-    }
-
-    scan->attack_box = &scan->attacker->h_cat->cat_box[0];
-    return scan->attack_box[1] != 0;
-}
-
-static bool prepare_catch_target(CatchScan* scan, s16 target_index) {
-    scan->target_index = target_index;
-
-    if (target_index == scan->attacker_index) {
-        return false;
-    }
-
-    if (hs[target_index].flag.results & 0x100) {
-        return false;
-    }
-
-    scan->target = q_hit_push[target_index];
-
-    if (scan->target->work_id != 1) {
-        return false;
-    }
-
-    scan->target_box = &scan->target->h_cau->cau_box[0];
-
-    if (scan->target_box[1] == 0) {
-        return false;
-    }
-
-    return !should_skip_catch_target(scan->attacker, scan->target);
-}
-
-static bool register_catch_hit(CatchScan* scan) {
-    if (!hit_check_subroutine(scan->attacker, scan->target, scan->attack_box, scan->target_box)) {
-        return false;
-    }
-
-    hs[scan->attacker_index].flag.results |= 0x1000;
-    hs[scan->attacker_index].my_hit = (u16)scan->target_index;
-    hs[scan->target_index].flag.results |= 0x100;
-    hs[scan->target_index].dm_me = (u16)scan->attacker_index;
-    scan->attacker->att_hit_ok = 0;
-    hs[scan->attacker_index].ah = scan->attack_box;
-    hs[scan->target_index].dh = scan->target_box;
-    scan->attacker->att_hit_ok = 0;
-    return true;
-}
-
-static void scan_catch_targets(CatchScan* scan) {
-    s16 si;
-
-    for (si = 0; si < hpq_in; si++) {
-        if (!prepare_catch_target(scan, si)) {
-            continue;
-        }
-
-        if (!register_catch_hit(scan)) {
-            continue;
-        }
-
-        break;
-    }
-}
-
-void catch_hit_check() { // 🟢
-    CatchScan scan;
-    s16 mi;
-
-    for (mi = 0; mi < hpq_in; mi++) {
-        if (!prepare_catch_attacker(&scan, mi)) {
-            continue;
-        }
-
-        scan_catch_targets(&scan);
-    }
-}
-
-static bool is_same_owner_target(WORK* mad, WORK* sad) {
-    if (mad->work_id == 1) {
-        if (sad->work_id == 1) {
-            return false;
-        }
-
-        if (((WORK_Other*)sad)->refrected != 0) {
-            return false;
-        }
-
-        return mad->id == ((WORK_Other*)sad)->master_id;
-    }
-
-    if (sad->work_id == 1) {
-        return ((WORK_Other*)mad)->master_id == sad->id;
-    }
-
-    return ((WORK_Other*)mad)->master_id == ((WORK_Other*)sad)->master_id;
-}
-
-static bool uses_vs_id_filter(WORK* mad, WORK* sad) {
-    if (!(mad->att.dipsw & 2)) {
-        return true;
-    }
-
-    if (sad->att.dipsw & 2) {
-        return false;
-    }
-
-    if (sad->work_id == 1) {
-        return true;
-    }
-
-    return !((WORK_Other*)sad)->refrected;
-}
-
-static bool can_vs_id_block_target(WORK* mad, WORK* sad) {
-    bool target_uses_dipswitch = sad->att.dipsw & 2;
-
-    if (mad->work_id == 1) {
-        return !target_uses_dipswitch;
-    }
-
-    if (mad->work_id == 8) {
-        return !target_uses_dipswitch;
-    }
-
-    return true;
-}
-
-static bool is_blocked_by_vs_id_filter(WORK* mad, WORK* sad) {
-    if (!uses_vs_id_filter(mad, sad)) {
-        return false;
-    }
-
-    if (!can_vs_id_block_target(mad, sad)) {
-        return false;
-    }
-
-    return !(mad->vs_id & sad->work_id);
-}
-
-typedef struct {
-    WORK* attacker;
-    WORK* target;
-    s16 attacker_index;
-    s16 target_index;
-} HitScan;
-
-static bool is_excluded_attack_damage_box_pair(s16 attack_box, s16 damage_box) {
-    return (attack_box == 2 || attack_box == 3) && (damage_box == 8 || damage_box == 9);
-}
-
-static bool is_damage_box_on_wrong_side(const HitScan* scan) {
-    if (((scan->attacker->rl_flag) + (scan->target->rl_flag)) & 1) {
-        return false;
-    }
-
-    if (scan->attacker->rl_flag) {
-        return !(scan->attacker->xyz[0].disp.pos <= scan->target->xyz[0].disp.pos);
-    }
-
-    return !(scan->attacker->xyz[0].disp.pos >= scan->target->xyz[0].disp.pos);
-}
-
-static bool is_auxiliary_damage_box_blocked(const HitScan* scan, s16 damage_box) {
-    return scan->attacker->att.dipsw & 4 && (damage_box >= 8 || scan->target->cg_ja.bhix == 0);
-}
-
-static bool should_skip_auxiliary_damage_box(const HitScan* scan, s16 damage_box) {
-    return ((damage_box > 3) && (damage_box < 0xA)) &&
-           (is_damage_box_on_wrong_side(scan) || is_auxiliary_damage_box_blocked(scan, damage_box));
-}
-
-static bool is_push_damage_box_blocked(const HitScan* scan, s16 damage_box) {
-    if (damage_box != 10) {
-        return false;
-    }
-
-    return !(scan->attacker->att.dipsw & 64) || scan->target->kind_of_waza & 0x60 || pcon_dp_flag ||
-           scan->target->pat_status == 0x26;
-}
-
-static bool should_end_damage_box_scan(const HitScan* scan, s16 damage_box) {
-    return damage_box > 3 && scan->attacker->att_hit_ok == 0;
-}
-
-static bool should_skip_damage_box(const HitScan* scan, s16 attack_box, s16 damage_box) {
-    return dmdat_adrs[damage_box][1] == 0 || is_excluded_attack_damage_box_pair(attack_box, damage_box) ||
-           should_skip_auxiliary_damage_box(scan, damage_box) || is_push_damage_box_blocked(scan, damage_box);
-}
-
-static void record_hit_if_stronger(const HitScan* scan, s16* attack_box_data, s16 attack_box, s16 damage_box) {
-    s16 overlap = hit_check_subroutine(scan->attacker, scan->target, attack_box_data, dmdat_adrs[damage_box]);
-
-    if (overlap > mkm_wk[scan->target_index]) {
-        hs[scan->attacker_index].flag.results |= 0x10;
-        hs[scan->attacker_index].my_hit = scan->target_index;
-        hs[scan->attacker_index].my_att = attack_box;
-        hs[scan->target_index].flag.results |= 1;
-        hs[scan->target_index].dm_me = scan->attacker_index;
-        hs[scan->target_index].dm_body = damage_box;
-        scan->attacker->att_hit_ok = 0;
-        mkm_wk[scan->target_index] = overlap;
-        hs[scan->attacker_index].ah = attack_box_data;
-        hs[scan->target_index].dh = dmdat_adrs[damage_box];
-    }
-}
-
-static bool check_attack_box_against_damage_boxes(const HitScan* scan, s16* mh, s16 lp) {
-    s16 lp2;
-
-    for (lp2 = 0; lp2 < 11; lp2++) {
-        if (should_end_damage_box_scan(scan, lp2)) {
-            return false;
-        }
-
-        if (should_skip_damage_box(scan, lp, lp2)) {
-            continue;
-        }
-
-        record_hit_if_stronger(scan, mh, lp, lp2);
-    }
-
-    return true;
-}
-
-static bool should_skip_attacker(WORK* attacker, WORK* target) {
-    return attacker->cg_ja.atix == 0 || attacker->att_hit_ok == 0 || is_blocked_by_vs_id_filter(attacker, target) ||
-           is_same_owner_target(attacker, target);
-}
-
-static void check_attacker_against_target(WORK* sad, s16 mi, s16 si) {
-    WORK* mad;
-    HitScan scan;
-    s16* mh;
-    s16 lp;
-    s16* assign2;
-
-    if (mi == si) {
-        return;
-    }
-
-    if (hs[mi].flag.results & 0x1110) {
-        return;
-    }
-
-    mad = q_hit_push[mi];
-
-    if (should_skip_attacker(mad, sad)) {
-        return;
-    }
-
-    scan.attacker = mad;
-    scan.target = sad;
-    scan.attacker_index = mi;
-    scan.target_index = si;
-    mh = &mad->h_att->att_box[0][0];
-
-    for (lp = 0; lp < 4; lp++, assign2 = mh += 4) {
-        if (mh[1] == 0) {
-            continue;
-        }
-
-        if (!check_attack_box_against_damage_boxes(&scan, mh, lp)) {
-            return;
-        }
-    }
-}
-
-void attack_hit_check() { // 🟢
-    WORK* sad;
-    s16* mh;
-    s16* sh;
-    s16 mi;
-    s16 si;
-    s16 lp;
-
-    s16* assign1;
-
-    for (si = 0; si < hpq_in; si++) {
-        if (hs[si].flag.results & 0x1101) {
-            continue;
-        }
-
-        sad = q_hit_push[si];
-        sh = sad->h_bod->body_dm[0];
-        mh = sad->h_han->hand_dm[0];
-
-        for (lp = 0; lp < 4; lp++, sh += 4, assign1 = mh += 4) {
-            dmdat_adrs[lp] = sh;
-            dmdat_adrs[lp + 4] = mh;
-        }
-
-        dmdat_adrs[8] = sad->h_att->att_box[2];
-        dmdat_adrs[9] = sad->h_att->att_box[3];
-        dmdat_adrs[10] = sad->h_hos->hos_box;
-
-        for (mi = 0; mi < hpq_in; mi++) {
-            check_attacker_against_target(sad, mi, si);
-        }
-    }
-}
-
-s16 hit_check_subroutine(WORK* wk1, WORK* wk2, const s16* hd1, const s16* hd2) { // 🟢
-    s16 d0;
-    s16 d1;
-    s16 d2;
-    s16 d3;
-
-    d0 = *hd1++;
-    d1 = *hd1++;
-
-    if (wk1->rl_flag) {
-        d0 = -d0;
-        d0 -= d1;
-    }
-
-    d0 += wk1->xyz[0].disp.pos;
-    d2 = *hd2++;
-    d3 = *hd2++;
-
-    if (wk2->rl_flag) {
-        d2 = -d2;
-        d2 -= d3;
-    }
-
-    d2 += wk2->xyz[0].disp.pos;
-    d2 += d3 - d0;
-    d3 += d1;
-
-    if ((u32)d2 >= d3) {
-        return 0;
-    }
-
-    d0 = (wk1->xyz[1].disp.pos + *hd1++) - (wk2->xyz[1].disp.pos + *hd2++);
-    d0 += d1 = *hd1;
-    d1 += *hd2;
-
-    if ((u32)d0 >= d1) {
-        return 0;
-    }
-
-    if (d2 > (d3 - d2)) {
-        d2 = d3 - d2;
-    }
-
-    return d2;
-}
-
-s32 hit_check_x_only(WORK* wk1, WORK* wk2, s16* hd1, s16* hd2) { // 🟢
-    s16 d0;
-    s16 d1;
-    s16 d2;
-    s16 d3;
-
-    d0 = *hd1++;
-    d1 = *hd1++;
-
-    if (wk1->rl_flag) {
-        d0 = -d0;
-        d0 -= d1;
-    }
-
-    d0 += wk1->xyz[0].disp.pos;
-    d2 = *hd2++;
-    d3 = *hd2++;
-
-    if (wk2->rl_flag) {
-        d2 = -d2;
-        d2 -= d3;
-    }
-
-    d2 += wk2->xyz[0].disp.pos;
-    d2 += d3 - d0;
-    d3 += d1;
-
-    if ((u32)d2 >= d3) {
-        return 0;
-    }
-
-    return 1;
-}
-
-void cal_hit_mark_position(WORK* wk1, WORK* wk2, s16* hd1, s16* hd2) { // 🟢
-    s16 d0 = *hd1++;
-    s16 d1 = *hd1++;
-    s16 d2;
-    s16 d3;
-
-    if (wk1->rl_flag) {
-        d0 = -d0;
-        d0 -= d1;
-    }
-
-    d0 += wk1->xyz[0].disp.pos;
-    d1 += d0;
-    d2 = *hd2++;
-    d3 = *hd2++;
-
-    if (wk2->rl_flag) {
-        d2 = -d2;
-        d2 -= d3;
-    }
-
-    d2 += wk2->xyz[0].disp.pos;
-    d3 += d2;
-
-    if (d0 < d2) {
-        d0 = d2;
-    }
-
-    if (d1 > d3) {
-        d1 = d3;
-    }
-
-    wk2->hit_mark_x = (d0 + d1) >> 1;
-
-    d0 = wk1->xyz[1].disp.pos + *hd1++;
-    d1 = *hd1 + d0;
-    d2 = wk2->xyz[1].disp.pos + *hd2++;
-    d3 = *hd2 + d2;
-
-    if (d0 < d2) {
-        d0 = d2;
-    }
-
-    if (d1 > d3) {
-        d1 = d3;
-    }
-
-    wk2->hit_mark_y = (d0 + d1) >> 1;
-}
-
-void get_target_att_position(WORK* wk, s16* tx, s16* ty) { // 🟢
-    s16 i;
-    s16(*ta)[4];
-
-    *tx = wk->xyz[0].disp.pos;
-    *ty = wk->xyz[1].disp.pos;
-    ta = &wk->h_att->att_box[0];
-
-    for (i = 0; i < 3; ta++, i++) {
-        if (!ta[0][0]) {
-            continue;
-        }
-
-        if (wk->rl_flag) {
-            *tx -= ta[0][0] + (ta[0][1] / 2);
-        } else {
-            *tx += ta[0][0] + (ta[0][1] / 2);
-        }
-
-        *ty += ta[0][2] + (ta[0][3] / 2);
-        break;
-    }
-}
-
-static s16 adjust_att_head_position(WORK* wk, s16 tx, s16 offset) {
-    s16 kx;
-
-    if (wk->rl_flag) {
-        kx = tx - offset;
-
-        if (tx < kx) {
-            return kx;
-        }
-
-        return tx;
-    }
-
-    kx = tx + offset;
-
-    if (tx > kx) {
-        return kx;
-    }
-
-    return tx;
-}
-
-s16 get_att_head_position(WORK* wk) { // 🟢
-    s16* ta;
-    s16 tx;
-    s16 i;
-
-    tx = wk->xyz[0].disp.pos;
-
-    if (wk->cg_ja.atix == 0) {
-        return tx;
-    }
-
-    ta = &wk->h_att->att_box[0][0];
-
-    for (i = 0; i < 3; i++) {
-        if (*ta) {
-            return adjust_att_head_position(wk, tx, *ta);
-        }
-
-        ta += 4;
-    }
-
-    return tx;
-}
-
-void hit_push_request(WORK* hpr_wk) { // 🟢
-    if (hpq_in < 31 && hpr_wk->cg_hit_ix != 0) {
-        q_hit_push[hpq_in++] = hpr_wk;
-    }
-}
-
-void clear_hit_queue() { // 🟢
-    s16 i;
-
-    hpq_in = 0;
-
-    for (i = 0; i < 32; i++) {
-        mkm_wk[i] = 0;
-    }
-
-    for (i = 0; i < 32; i++) {
-        q_hit_push[i] = 0;
-    }
-
-    SDL_zeroa(hs);
-}
-
-static bool is_reversed_gill(PLW* as) {
-    if (as->wu.work_id != 1) {
-        return false;
-    }
-
-    if (as->player_number != CHAR_GILL) {
-        return false;
-    }
-
-    return as->wu.rl_flag;
-}
-
-static u16 apply_flame_damage_attribute(PLW* as, u16 ix) {
-    if (is_reversed_gill(as)) {
-        ix = attr_freeze_tbl[ix - 32];
-        as->wu.at_attribute = 3;
-        return ix;
-    }
-
-    return attr_flame_tbl[ix - 32];
-}
-
-static u16 apply_freeze_damage_attribute(PLW* as, u16 ix) {
-    if (is_reversed_gill(as)) {
-        ix = attr_flame_tbl[ix - 32];
-        as->wu.at_attribute = 1;
-        return ix;
-    }
-
-    return attr_freeze_tbl[ix - 32];
-}
-
-s16 change_damage_attribute(PLW* as, u16 atr, u16 ix) { // 🟢
-    switch (atr) {
-    case 1:
-        ix = apply_flame_damage_attribute(as, ix);
-        break;
-
-    case 2:
-        ix = attr_thunder_tbl[ix - 32];
-        break;
-
-    case 3:
-        ix = apply_freeze_damage_attribute(as, ix);
-        break;
-    }
-
-    return ix;
-}
-
-s16 get_sky_nm_damage(u16 ix) { // 🟢
-    ix -= 32;
-    return sky_nm_damage_tbl[ix];
-}
-
-s16 get_sky_sp_damage(u16 ix) { // 🟢
-    ix -= 32;
-    return sky_sp_damage_tbl[ix];
-}
-
-s16 get_kagami_damage(u16 ix) { // 🟢
-    ix -= 32;
-    return kagami_damage_tbl[ix];
-}
-
-s16 get_grd_hand_damage(u16 ix) { // 🟢
-    ix -= 32;
-    return grd_hand_damage_tbl[ix];
-}
-
-u8 check_head_damage(s16 ix) { // 🟢
-    ix -= 32;
-    return hddm_damage_tbl[ix];
-}
-
-u8 check_trunk_damage(s16 ix) { // 🟢
-    ix -= 32;
-    return trdm_damage_tbl[ix];
-}
-
-u8 check_ttk_damage_request(s16 ix) { // 🟢
-    ix -= 32;
-    return ttk_damage_req_tbl[ix];
-}
 
 const u16 chain_normal_ground_table[8] = { 0x760, 0x660, 0x640, 0x440, 0x400, 0x0, 0x0, 0x0 };
 const u16 chain_hidou_nm_ground_table[8] = { 0x770, 0x770, 0x740, 0x470, 0x600, 0x60, 0x0, 0x0 };
